+++ workspaces/core/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @heathmont/moon-core-tw
 
+## 9.2.1
+
+### Patch Changes
+
+- fixes
+- Updated dependencies [undefined]
+  - @heathmont/moon-icons-tw@9.2.1
+
 ## 9.2.0
 
 ### Minor Changes